Revert "Revert "Lambda関数を分割""
