--- api/resources/callback/struct.go
@@ -1,10 +1,0 @@
-package callback
-
-type tokenResponse struct {
-	AccessToken  string `json:"access_token"`
-	TokenType    string `json:"token_type"`
-	RefreshToken string `json:"refresh_token"`
-	ExpiresIn    int    `json:"expires_in"`
-	Scope        string `json:"scope"`
-	IdToken      string `json:"id_token"`
-}
