+++ functions/authorizer/verify.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"strings"
-	"ubic-food/api/token"
+	"ubic-food/tools/token"
 )
 
 func verify(authZHeader string) (token.Payload, error) {
