+++ functions/image_post/main.go
@@ -1,32 +1,32 @@
-package image
+package main
 
 import (
 	"bytes"
 	"encoding/base64"
 	"encoding/json"
-	"ubic-food/api/dynamodb"
-	"ubic-food/api/response"
-	"ubic-food/api/token"
+	"ubic-food/tools/dynamodb"
+	"ubic-food/tools/response"
 
-	"ubic-food/api/s3"
+	"ubic-food/tools/s3"
 
 	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-lambda-go/lambda"
 )
 
 type imagePostResponse struct {
 	ImageUrl string `json:"imageUrl"`
 }
 
-func ImagePost(request events.APIGatewayProxyRequest, idTokenPayload token.Payload) events.APIGatewayProxyResponse {
+func handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
 	reqBody := request.Body
 	dec, err := base64.StdEncoding.DecodeString(reqBody)
 	if err != nil {
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 
 	fileName, err := dynamodb.GenerateID()
 	if err != nil {
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 	fileName += ".jpeg"
 
@@ -35,16 +35,19 @@ func ImagePost(request events.APIGatewayProxyRequest, idTokenPayload token.Paylo
 
 	location, err := s3.Upload(&buf, fileName, request.Headers["Content-Type"])
 	if err != nil {
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 
 	res, err := json.Marshal(imagePostResponse{
 		ImageUrl: location,
 	})
 	if err != nil {
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 
-	return response.StatusCode200(string(res))
+	return response.StatusCode200(string(res)), nil
+}
 
+func main() {
+	lambda.Start(handler)
 }
