+++ functions/transaction_post/main.go
@@ -1,36 +1,43 @@
-package transaction
+package main
 
 import (
 	"encoding/json"
 	"errors"
 	"fmt"
 	"time"
-	"ubic-food/api/dynamodb"
-	"ubic-food/api/response"
-	"ubic-food/api/token"
-	"ubic-food/api/types"
+
+	"ubic-food/tools/dynamodb"
+	"ubic-food/tools/response"
+	"ubic-food/tools/token"
+	"ubic-food/tools/types"
 
 	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-lambda-go/lambda"
 )
 
 type requestBody struct {
 	FoodId string `json:"foodId"`
 }
 
-func TransactionPost(request events.APIGatewayProxyRequest, idTokenPayload token.Payload) events.APIGatewayProxyResponse {
+func handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
+
+	idTokenPayload, err := token.GetIdTokenPayloadByRequest(request)
+	if err != nil {
+		return response.StatusCode500(err), nil
+	}
 
 	bodyStr := request.Body
 	var body requestBody
-	err := json.Unmarshal([]byte(bodyStr), &body)
+	err = json.Unmarshal([]byte(bodyStr), &body)
 	if err != nil {
-		return response.StatusCode400(err)
+		return response.StatusCode400(err), nil
 	}
 
 	var transaction types.Transaction
 	transaction.ID, err = dynamodb.GenerateID()
 	if err != nil {
 		fmt.Println("Error generating ID:", err.Error())
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 	transaction.Date = int(time.Now().Unix())
 	food := types.Food{
@@ -41,12 +48,12 @@ func TransactionPost(request events.APIGatewayProxyRequest, idTokenPayload token
 	err = food.Get()
 	if err != nil {
 		fmt.Println("Error getting food:", err.Error())
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 	transaction.Food = food
 
 	if food.Stock < 1 {
-		return response.StatusCode400(errors.New("food is out of stock"))
+		return response.StatusCode400(errors.New("food is out of stock")), nil
 	}
 
 	/* err = dynamodb.AddIntData(-1, food.ID, "food-stock")
@@ -61,16 +68,20 @@ func TransactionPost(request events.APIGatewayProxyRequest, idTokenPayload token
 		err2 := dynamodb.AddIntData(1, food.ID, "food-stock")
 		if err2 != nil {
 			fmt.Println("Error adding food stock:", err2.Error())
-			return response.StatusCode500(err2)
+			return response.StatusCode500(err2), nil
 		}
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 
 	resBody, err := json.Marshal(transaction)
 	if err != nil {
 		fmt.Println("Error marshalling transaction:", err.Error())
-		return response.StatusCode500(err)
+		return response.StatusCode500(err), nil
 	}
 
-	return response.StatusCode200(string(resBody))
+	return response.StatusCode200(string(resBody)), nil
+}
+
+func main() {
+	lambda.Start(handler)
 }
