+++ tools/dynamodb/struct.go
