+++ tools/hash/hash.go
