+++ tools/random/random.go
