+++ tools/token/struct.go
