+++ tools/types/food.go
@@ -1,7 +1,7 @@
 package types
 
 import (
-	"ubic-food/api/dynamodb"
+	"ubic-food/tools/dynamodb"
 )
 
 type Food struct {
